Only shift cycles
Only cycles can be shifted. Shifting a path that is not a cycle, introduces
a new edge that does not actually exist. For example, [A1, B1, C1, A2] is a
valid path that is not a cycle. There is no edge between A1 and A2 since they
are in the same group.

Shifting this in any direction and moving A1 and A2 out of their current first
and last positions introduces an edge between A1 and A2. e.g. [B1, C1, A2, A1]
or [A2, A1, B1, C1].

If a path is a cycle, then that means there does exist an edge between the
first and last nodes of the path. Therefore, shifting a cycle does not
introduce any new edges.